Update index.html
Capitalised spelling of "pyglet" -> "Pyglet" as it is the name of the project. Also added a button that takes you straight to the documentation from the homepage, below the GitHub button.

I don't generally work in HTML so this might not be correct, but thought I might give it a go! :)